fixed bug that emerged when moving from node to deno
after switching from node to deno, parsing player_response via `new URLSearchParams(text)` would not decode +s into spaces in the JSON string.

I'm not really sure if it's even supposed to do that in the first place; the documentation for `URLSearchParams` is kinda vague?

but it was kinda a hacked-together solution in the first place, so I replaced all `scurl` calls with the appropriate `fetch` calls, and everything works now.